finalizing

+++ src/commonMain/kotlin/dev/voir/anyexchange/sdk/dto/CurrencyRateWithDateDto.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CurrencyRateWithDateDto(
     val code: String,
-    val rate: Float,
+    val rate: Double,
     val source: String,
     val date: String,
-    val fluctuation: Float
+    val fluctuation: Double
 )
